Add volunteer map to home page
