+++ app/javascript/packs/components/ScrollToTopOnMount.jsx
@@ -1,11 +1,13 @@
-import React, { useEffect } from 'react';
+import React, { useEffect } from "react";
 
 const ScrollToTopOnMount = () => {
   useEffect(() => {
-    window.scrollTo(0, 0);
+    setTimeout(() => {
+      window.scrollTo(0, 0);
+    });
   }, []);
 
   return null;
-}
+};
 
 export default ScrollToTopOnMount;
