+++ app/javascript/packs/pages/home/Home.jsx
@@ -5,12 +5,14 @@ import VolunteerCallout from "./VolunteerCallout";
 import HelpCallout from "./HelpCallout";
 import About from "./About";
 import Contributors from "./Contributors";
+import Volunteers from "./Volunteers";
 
 const Home = () => (
   <>
     <Hero />
     <SplitSection />
     <VolunteerCallout />
+    <Volunteers />
     <HelpCallout />
     <About />
     <Contributors />
